simple black jack game
Once the players turn is over the dealer will flip over their face down card and hit or stay according to a set of rules.

If their hand has a value of 17 or more they must stay.
If their hand has a value of 16 or less they must hit.
The dealer must continue to hit until their hand has a value of 17 or greater, if the value of their hand ever exceeds 21, they immediately lose.
Once the dealer stays or busts the hand/round is over and the winner is decided by the player with the hand that has the highest value not exceeding 21.

If the player is the winner their bet is doubled and returned to them.
If the dealer is the winner they take the players bet.
If the player and dealer tie, then the original bet is returned to the player.
In addition, a special rule in Blackjack is that if a player is dealt a hand with a value of 21 this is considered a natural or "blackjack". If the player has a natural and the dealer does not the dealer pays the player one and a half times their original bet and the hand is immediately over.